fixing the schema and jpa type missmatch and getting on stable version

+++ src/main/resources/db/migration/V4__create_meter_reading.sql
@@ -2,8 +2,8 @@ CREATE TABLE meter_reading (
     id UUID PRIMARY KEY,
     tenant_id UUID NOT NULL,
     reading_month DATE NOT NULL,
-    previous_reading NUMERIC(10,2) NOT NULL,
-    current_reading NUMERIC(10,2) NOT NULL,
+    previous_reading BIGINT NOT NULL,
+    current_reading BIGINT NOT NULL,
     units_consumed NUMERIC(10,2) NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
